Cleanup scripts even further
Rename script-called scripts to start with an underscore so that
they are sorted first in "printenv" output.  For long user-called
scripts, rename to start with an underscore and make the main
script just run that script.

Now that there is the "nand write.auto" feature, modify scripts to
use this feature to write out the nand images.  As a result,
burnmmcxloader is completely removed.

Convert more scripts to the newline format to make them easier to
read from u-boot's "printenv"

git-svn-id: svn+ssh://svn/srv/svn/eps_svn/software/products/linux/u-boot/u-boot-2011.06@29229 d7e73fac-2431-0410-abd5-fd325483733a